requirements.txt
pyperclip >=1.8.2
pycryptodome >=3.19.0